% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grouped_mean}
\alias{grouped_mean}
\title{Mean by groups of columns}
\usage{
grouped_mean(pBigMat, group_cols, summ_col = 4L, zeros = FALSE)
}
\arguments{
\item{pBigMat}{`big.matrix` object pointer (the `@address` slot!).}

\item{group_cols}{Vector of columns along which means will be calculated.}

\item{summ_col}{Column to summarize.}

\item{zeros}{Boolean for whether to compute the proportion of zeros rather than
the mean. Defaults to `FALSE`.}
}
\description{
Mean by groups of columns
}
