% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cages.R
\name{sim_cages}
\alias{sim_cages}
\title{Simulate multiple cages and simplify output.}
\usage{
sim_cages(n_cages, N_0, max_t, R, A, D_binom, D_nb, process_error, plant_mort_0,
  plant_mort_1, plant_death_age_mean, plant_death_age_sd, repl_times, repl_age,
  extinct_N, n_cores = 1, by_cage = FALSE, condense = TRUE,
  show_progress = FALSE)
}
\arguments{
\item{n_cages}{Number of cages to simulate.}

\item{N_0}{Starting abundances for each aphid line on each plant.}

\item{max_t}{Max time points to simulate for each cage.}

\item{R}{Growth rates for each line.}

\item{A}{Density dependence for each line.}

\item{D_binom}{Data frame containing intercept and coefficient estimates for
binomial GLM of `<whether dispersal > 0> ~ N_t + N_t^2`.
There should be one set of estimates for each line regardless of whether the model
estimated separate values for each line.}

\item{D_nb}{Data frame containing intercept and overdispersion estimates for
negative binomial GLM of `<# dispersed | dispersal occurs> ~ 1`.
There should be one set of estimates for each line regardless of whether the model
estimated separate values for each line.}

\item{process_error}{SD of process error.}

\item{plant_mort_0}{Intercept for plant-death-induced aphid mortality, one per line.}

\item{plant_mort_1}{Coefficient for plant-death-induced aphid mortality, one per line.}

\item{plant_death_age_mean}{Mean of distribution of plant-death ages.}

\item{plant_death_age_sd}{SD of distribution of plant-death ages.}

\item{repl_times}{Vector of dates on which to replace plants. Use 0-based indexing!}

\item{repl_age}{Number of days past death when a plant will be replaced.}

\item{n_cores}{Number of cores to use. Defaults to \code{1}.}

\item{show_progress}{Boolean for whether to show progress bar. Defaults to
\code{FALSE}.}
}
\description{
Simulate multiple cages and simplify output.
}
