% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_cages.R
\name{sim_cages}
\alias{sim_cages}
\title{Simulate multiple cages and simplify output.}
\usage{
sim_cages(n_cages, N_0, max_t, R, A, D_0, D_1, process_error, plant_mort_coefs,
  plant_death_age, repl_times, repl_plants, n_cores = 1,
  show_progress = FALSE)
}
\arguments{
\item{n_cages}{Number of cages to simulate.}

\item{N_0}{Starting abundances for each aphid line on each plant.}

\item{max_t}{Max time points to simulate for each cage.}

\item{R}{Growth rates for each line.}

\item{A}{Density dependence for each line.}

\item{D_0}{Intercept for regression of dispersal ~ # aphids, one for each line.}

\item{D_1}{Coefficient for regression of dispersal ~ # aphids, one for each line.}

\item{process_error}{SD of process error.}

\item{plant_mort_coefs}{Coefficients for plant-death-induced aphid mortality.}

\item{plant_death_age}{Date on which plants begin to die. Use 0-based indexing!}

\item{repl_times}{Vector of dates on which to replace plants. Use 0-based indexing!}

\item{repl_plants}{List of vectors, each vector indicating the plants to replace
on the corresponding date in `repl_times`. Use 0-based indexing!}

\item{n_cores}{Number of cores to use. Defaults to \code{1}.}

\item{show_progress}{Boolean for whether to show progress bar. Defaults to
\code{FALSE}.}
}
\description{
Simulate multiple cages and simplify output.
}
